common.hh: fixed type error with splitext
darcs-hash:20100301200300-21d25-b198f44fd69e15ebcc07df945c470e5af99a5f0f.gz

+++ common.hh
@@ -85,8 +85,8 @@ void arange(blitz::Array<T,1> &output, int start, int stop, int step=1)
  * @return split point
  */
 inline
-unsigned int splitext(const std::string &in, std::string &froot, std::string &ext) {
-	unsigned int fidx = in.rfind(".");
+size_t splitext(const std::string &in, std::string &froot, std::string &ext) {
+	size_t fidx = in.rfind(".");
 	if (fidx==std::string::npos)
 		return -1;
 	froot.assign(in.substr(0,fidx));
